Add context to the testing section of the README
This adds some context and resources to the testing section of the README.  The project doesn't really have testing, but as a group we all felt more comfortable making a habit of TDD.  I've added some links to resources. While most people will probably and call it a day, perhaps the links will save another dev some time down the road.